Adds sample user component. Explanation of hr procedure below
- Explanation of hr adding: Because we don't want an hr before
the scene tree, for the permanent elements we add hr's *after*
each element, except for the last one. If there are no user
components, we don't want an hr there. For the user components,
we had hr *before* each component. That way the first user
component has an hr between it and the permanent elements above,
and the last one doesn't have any after it.
- Gets rid of material reference html
- TODO: get rid of console.log in toggle collapse